feat: upgrade deps, go version (#17)
the goal of this PR is simply to get the project compiling on main
again. the ended up being an involved process because the dependencies
had mismatches of their url and go mod path. that is now all addressed.
the new release should create a new version in the catalystcommunity
image repository, which should fix issues with the helm chart pointing
to a missing image.

- upgrades all of the catalystcommunity dependencies so that the project
  will compile on main again, now that the dependencies have changed
  location.
- upgrade grpc dependencies as the protos had their plugins upgraded to
  get around 404 problems with the buf plugins.
- upgrade go, as the new buf plugin dependencies now require a higher go
  version.